requirements.txt
python-dotenv
beautifulsoup4
selenium
Pillow
sendgrid==5.6.0